Updated "Show Photo" screen to ask for location permission and retrieve the current location data (latitude/longitude) from photos taken in the app.

+++ app/src/main/java/nz/ac/canterbury/guessit/Photo.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "photo")
 class Photo(
     @PrimaryKey @ColumnInfo val file: String,
-    @ColumnInfo val latitude: Float,
-    @ColumnInfo val longitude: Float)
+    @ColumnInfo val latitude: Double,
+    @ColumnInfo val longitude: Double)
